@@ -1,13 +1,15 @@
-<?php
+<?php declare(strict_types=1);
+
+namespace Lisachenko\Protocol;
+
 /**
+ * FCGI constants.
+ *
  * @author Alexander.Lisachenko
- * @date 14.07.2014
  */
-
-namespace Protocol;
-
 class FCGI
 {
+
     /**
      * Number of bytes in a FCGI_Header.  Future versions of the protocol
      * will not reduce this number.
@@ -27,17 +29,18 @@ class FCGI
     /**
      * Values for type component of FCGI_Header
      */
-    const BEGIN_REQUEST      = 1;
-    const ABORT_REQUEST      = 2;
-    const END_REQUEST        = 3;
-    const PARAMS             = 4;
-    const STDIN              = 5;
-    const STDOUT             = 6;
-    const STDERR             = 7;
-    const DATA               = 8;
-    const GET_VALUES         = 9;
-    const GET_VALUES_RESULT  = 10;
-    const UNKNOWN_TYPE       = 11;
+    const
+        BEGIN_REQUEST     = 1,
+        ABORT_REQUEST     = 2,
+        END_REQUEST       = 3,
+        PARAMS            = 4,
+        STDIN             = 5,
+        STDOUT            = 6,
+        STDERR            = 7,
+        DATA              = 8,
+        GET_VALUES        = 9,
+        GET_VALUES_RESULT = 10,
+        UNKNOWN_TYPE      = 11;
 
     /**
      * Value for requestId component of FCGI_Header
@@ -52,15 +55,18 @@ class FCGI
     /**
      * Values for role component of FCGI_BeginRequestBody
      */
-    const RESPONDER  = 1;
-    const AUTHORIZER = 2;
-    const FILTER     = 3;
+    const
+        RESPONDER  = 1,
+        AUTHORIZER = 2,
+        FILTER     = 3;
 
     /**
      * Values for protocolStatus component of FCGI_EndRequestBody
      */
-    const REQUEST_COMPLETE = 0;
-    const CANT_MPX_CONN    = 1;
-    const OVERLOADED       = 2;
-    const UNKNOWN_ROLE     = 3;
+    const
+        REQUEST_COMPLETE = 0,
+        CANT_MPX_CONN    = 1,
+        OVERLOADED       = 2,
+        UNKNOWN_ROLE     = 3;
+
 }
