@@ -1,24 +1,23 @@
-<?php
-/**
- * @author Alexander.Lisachenko
- * @date 14.07.2014
- */
+<?php declare(strict_types=1);
 
-namespace Protocol\FCGI\Record;
+namespace Lisachenko\Protocol\FCGI\Record;
 
-use Protocol\FCGI;
-use Protocol\FCGI\Record;
+use Lisachenko\Protocol\FCGI;
+use Lisachenko\Protocol\FCGI\Record;
 
 /**
  * The application sends a FCGI_END_REQUEST record to terminate a request, either because the application
  * has processed the request or because the application has rejected the request.
+ *
+ * @author Alexander.Lisachenko
  */
 class EndRequest extends Record
 {
+
     /**
      * The appStatus component is an application-level status code. Each role documents its usage of appStatus.
      *
-     * @var integer
+     * @var int
      */
     protected $appStatus = 0;
 
@@ -35,8 +34,7 @@ class EndRequest extends Record
      *   FCGI_UNKNOWN_ROLE: rejecting a new request.
      *      This happens when the Web server has specified a role that is unknown to the application.
      *
-     *
-     * @var integer
+     * @var int
      */
     protected $protocolStatus = FCGI::REQUEST_COMPLETE;
 
@@ -47,12 +45,12 @@ class EndRequest extends Record
      */
     protected $reserved1;
 
-    public function __construct($protocolStatus = FCGI::REQUEST_COMPLETE, $appStatus = 0, $reserved = '')
+    public function __construct(int $protocolStatus = FCGI::REQUEST_COMPLETE, int $appStatus = 0, string $reserved = '')
     {
-        $this->type           = FCGI::END_REQUEST;
+        $this->type = FCGI::END_REQUEST;
         $this->protocolStatus = $protocolStatus;
-        $this->appStatus      = $appStatus;
-        $this->reserved1      = $reserved;
+        $this->appStatus = $appStatus;
+        $this->reserved1 = $reserved;
         $this->setContentData($this->packPayload());
     }
 
@@ -60,10 +58,8 @@ class EndRequest extends Record
      * Returns app status
      *
      * The appStatus component is an application-level status code. Each role documents its usage of appStatus.
-     *
-     * @return int
      */
-    public function getAppStatus()
+    public function getAppStatus(): int
     {
         return $this->appStatus;
     }
@@ -80,39 +76,27 @@ class EndRequest extends Record
      *      This happens when the application runs out of some resource, e.g. database connections.
      *   FCGI_UNKNOWN_ROLE: rejecting a new request.
      *      This happens when the Web server has specified a role that is unknown to the application.
-     *
-     * @return int
      */
-    public function getProtocolStatus()
+    public function getProtocolStatus(): int
     {
         return $this->protocolStatus;
     }
 
     /**
-     * Method to unpack the payload for the record
-     *
-     * @param Record|self $self Instance of current frame
-     * @param string $data Binary data
-     *
-     * @return Record
+     * {@inheritdoc}
+     * @param static $self
      */
-    protected static function unpackPayload(Record $self, $data)
+    protected static function unpackPayload($self, string $data): void
     {
-        list(
+        [
             $self->appStatus,
             $self->protocolStatus,
             $self->reserved1
-        ) = array_values(unpack("NappStatus/CprotocolStatus/a3reserved", $data));
-
-        return $self;
+        ] = array_values(unpack("NappStatus/CprotocolStatus/a3reserved", $data));
     }
 
-    /**
-     * Implementation of packing the payload
-     *
-     * @return string
-     */
-    protected function packPayload()
+    /** {@inheritdoc} */
+    protected function packPayload(): string
     {
         return pack(
             "NCa3",
@@ -121,4 +105,5 @@ class EndRequest extends Record
             $this->reserved1
         );
     }
+
 }
