@@ -1,26 +1,26 @@
-<?php
+<?php declare(strict_types=1);
+
+namespace Lisachenko\Protocol\FCGI\Record;
+
+use PHPUnit\Framework\TestCase;
+use Lisachenko\Protocol\FCGI;
+
 /**
  * @author Alexander.Lisachenko
- * @date   23.10.2015
  */
-
-namespace Protocol\FCGI\Record;
-
-use Protocol\FCGI;
-
-class StdinTest extends \PHPUnit_Framework_TestCase
+class StdinTest extends TestCase
 {
     protected static $rawMessage = '01050000000404007465737400000000';
 
-    public function testPacking()
+    public function testPacking(): void
     {
         $request = new Stdin('test');
         $this->assertEquals($request->getContentData(), 'test');
         $this->assertEquals($request->getType(), FCGI::STDIN);
-        $this->assertSame(self::$rawMessage, bin2hex($request));
+        $this->assertSame(self::$rawMessage, bin2hex((string) $request));
     }
 
-    public function testUnpacking()
+    public function testUnpacking(): void
     {
         $request = Stdin::unpack(hex2bin(self::$rawMessage));
         $this->assertEquals($request->getType(), FCGI::STDIN);
